Rename Email.java to Detail.java and update Detail.java to model task detail
